Move consent-toggle component to components folder
